started contrib directory, slice moved into it
I'm probably going to leave this alone now, but I'll keep this branch
around since it has the namespace setup for a contrib directory.
