+++ contrib/contrib.ts
@@ -1,0 +1,1 @@
+export namespace Contrib {}
