+++ package.json
@@ -2,8 +2,8 @@
   "name": "@sodaviz/soda",
   "version": "0.7.0",
   "description": "A library for visualizing biological annotation.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
   "prepublish": "make build",
   "scripts": {
     "test": "test.sh"
